Refactor + create securitypolicy for targetnetworkobject
